% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/anova.R
\name{all_factor_anova}
\alias{all_factor_anova}
\title{Performs ANOVA for all factors.}
\usage{
all_factor_anova()
}
\description{
ANOVA for all single factors are calculated.
The reported p-values are tghe unadjusted p-values (no multiple testing
correction)
}

