% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_reshape.R
\name{bdl_fun_data}
\alias{bdl_fun_data}
\title{Calculate function fun over time points.}
\usage{
bdl_fun_data(data, samples, fun)
}
\description{
The function is used to calculate the mean and sd on the BDL data set.
}

