% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_reshape.R
\name{bdl_matrix_data}
\alias{bdl_matrix_data}
\title{Transform BDL data frame into list of matrices.}
\usage{
bdl_matrix_data(data, samples, Nrepeats = 5)
}
\description{
Transform BDL data frame into list of matrices.
}

