% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_reshape.R
\name{bdl_mean_data}
\alias{bdl_mean_data}
\title{Calculate mean data average over time points.}
\usage{
bdl_mean_data(data, samples)
}
\description{
Use the time point repeats for averaging. The time points are defined
in the sample definitions
}

