% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_reshape.R
\name{bdl_sd_data}
\alias{bdl_sd_data}
\title{Calculate sd data over time points.}
\usage{
bdl_sd_data(data, samples)
}
\description{
Use the time point repeats for calculation of sd. The time points are defined
in the sample definitions
}

