\name{BDLanalysis-package}
\alias{BDLanalysis-package}
\alias{BDLanalysis}
\docType{package}
\title{
Statistical analysis for data sets of bile duct ligated (BDL) mice.
}
\description{
Complete statistical analysis and data sets accompanying the publication Pathobiochemical signatures of cholestatic liver disease in bile duct ligated mice. Pathobiochemical signatures of cholestatic liver disease in bile duct ligated (BDL) mice were analysed 
in a comprehensive data set of serum markers, histological parameters and transcript profiles at 8 time points after bile duct ligation in mice, comprising different stages of the disease (> 6000 data points).
}
\details{
\tabular{ll}{
Package: \tab BDLanalysis\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-08-03\cr
License: \tab GPLv3\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Matthias Koenig

Maintainer: Matthias Koenig <konigmatt@googlemail.com>
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
