% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BDLanalysis.R
\docType{data}
\name{BDLanova}
\alias{BDLanova}
\title{ANOVA results.}
\format{A data frame with with 154 rows and 5 variable providing p-values of ANOVA}
\description{
Table of ANOVA results for filtering.
}

