% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BDLanalysis.R
\docType{data}
\name{BDLfactors}
\alias{BDLfactors}
\title{Factor description for BDLdata.}
\format{A data frame with 154 rows providing additional information for the samples.}
\description{
Factor description for the BDLdata factor data set. Here the factor ids, description and type are defined.
}

