% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BDLanalysis.R
\docType{data}
\name{BDLprobes}
\alias{BDLprobes}
\title{Fluidigm probe mapping to UniProt.}
\format{A data frame with with 40 rows and 9 variable providing additional information for the samples: \code{time}, \code{time_fac}, \code{time_point}, \code{repeats} ...).}
\description{
Maps the gene identifier to UniProt identifiers for additional information
retrieval. This information is used for customizing the plots.
}

