% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BDLsamples}
\alias{BDLsamples}
\title{Sample description for BDL mice corresponding to BDLdata.}
\format{A data frame with factor variables: \code{test}, \code{sex},
  \code{name}, \code{n} and \code{prop} (\code{n} divided by total number
  of applicants in that year).}
\usage{
BDLsamples
}
\description{
Full data of all factors.
}
\keyword{datasets}

