% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{f_cor_pair_plot}
\alias{f_cor_pair_plot}
\title{Correlation plot between two factors}
\usage{
f_cor_pair_plot(name_A, name_B, single_plots = TRUE)
}
\description{
Correlation plot between two factors
}

