% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{HeatmapColors}
\alias{HeatmapColors}
\title{Heatmap color ramp function.}
\usage{
HeatmapColors()
}
\description{
Definition of heatmap colors via a colorRampPalette.
The returned function can be called with the numbers of colors in the heatmap.
}

