% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_all_factors}
\alias{plot_all_factors}
\title{Creates single factor plots of all factors.}
\usage{
plot_all_factors(path)
}
\description{
Iterates over all factor variables in the BDLdata frame and creates
the individual factor plots.
}

