% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_cor_pair}
\alias{plot_cor_pair}
\title{Correlation plot between two factors.}
\usage{
plot_cor_pair(name_A, name_B, single_plots = TRUE)
}
\description{
Plot two individual factors and correlation scatter plot.
}

