% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/anova.R
\name{significant_code}
\alias{significant_code}
\title{Create significant code for given p-value.}
\usage{
significant_code(p.value)
}
\description{
Function returns a significance string for the
given p-value similar to the internal R functions.
}

