% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/anova.R
\name{single_factor_anova}
\alias{single_factor_anova}
\title{Perform anova for given factor.}
\usage{
single_factor_anova(name)
}
\description{
Similar implementation to
http://www.r-tutor.com/elementary-statistics/analysis-variance/completely-randomized-design
}

