% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ys1_yr1.R
\name{son.fA}
\alias{son.fA}
\title{Relative count of number of identical slopes between two time courses.}
\usage{
son.fA(a, b, time_pts)
}
\description{
Relative count of number of identical slopes between two time courses.
}

