% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ys1_yr1.R
\name{son.fA.timecourse}
\alias{son.fA.timecourse}
\title{fA calculation for time courses with multiple repeats.}
\usage{
son.fA.timecourse(a, b, time_pts)
}
\description{
fA with multiple repeats i.e. multiple samples measured for the same time course.
a and b are matrices of the form [Nt, Nr], with
  Nt : number of time points
  Nr : number of repeats == number of samples per condition in case of time course
This is only applicable if the individual measurements within a column are from the
same sample !, i.e. real timecourse repeats. If every timepoint within a repeat
belongs to a different sample the timepoint version must be used.
}

