% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ys1_yr1.R
\name{son.fA.timepoints}
\alias{son.fA.timepoints}
\title{fA caluclation for time points with multiple repeats.}
\usage{
son.fA.timepoints(a, b, time_pts)
}
\description{
Calculate all the permutations of the datapoint slopes,
and count the identical ones.
}

