% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ys1_yr1.R
\name{son.fA_star}
\alias{son.fA_star}
\title{Adapted slope measurement.}
\usage{
son.fA_star(a, b, time_pts)
}
\description{
Since the concordance index A counts only the number of time intervals
in which there is an agreement in the sign of the change between two
profiles, it may loose the information of the size of the change.
Use the size of the change and compute the correlation coefficient between the
differences of the profiles in the time intervals.
This works on the distances, not the slopes! Which is different if the time points
are not equidistant.
}

