% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ys1_yr1.R
\name{son.fA_star2}
\alias{son.fA_star2}
\title{Adapted slope measurement (spearman)}
\usage{
son.fA_star2(a, b, time_pts)
}
\description{
Similar to fA_star but uses spearman for slope correlation calculation.
}

