% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ys1_yr1.R
\name{son.fA_star3}
\alias{son.fA_star3}
\title{Calculating the distance on the slopes.}
\usage{
son.fA_star3(a, b, time_pts)
}
\description{
The distance based measurement should be preferred in case of unequal time points.
}

