% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ys1_yr1.R
\name{son.fM_star}
\alias{son.fM_star}
\title{Alternative index to M.}
\usage{
son.fM_star(a, b, time_pts)
}
\description{
M index utilizing the distances between two profiles time
points where the max/min is attained.
Uses the difference in indices !
}

