% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ys1_yr1.R
\name{son.fM_star2}
\alias{son.fM_star2}
\title{Alternative index to M.}
\usage{
son.fM_star2(a, b, time_pts)
}
\description{
Alternative index to M which utilizes the distances between two profiles time.
points where the max/min is attained.
Uses the difference in times
}

