% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ys1_yr1.R
\name{son.fM_star3}
\alias{son.fM_star3}
\title{Curvature based M measure.}
\usage{
son.fM_star3(a, b, time_pts)
}
\description{
Contains implicitly the information about minimum and maximum
of a timecourse.
}

