% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ys1_yr1.R
\name{son.fR_star.matrix}
\alias{son.fR_star.matrix}
\title{fR_star calculation for matrix with multiple repeats.}
\usage{
son.fR_star.matrix(a, b)
}
\description{
Works for timecourse and timepoint repeat data.
}

