% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ys1_yr1.R
\name{son.fS_star.matrix}
\alias{son.fS_star.matrix}
\title{fS_star calculation for matrix with multiple repeats.}
\usage{
son.fS_star.matrix(a, b)
}
\description{
Works in case of the timecourse and timepoint repeat data.
}

