% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ys1_yr1.R
\name{son.fS_star}
\alias{son.fS_star}
\title{Adapted Spearman correlation S*.}
\usage{
son.fS_star(a, b)
}
\description{
Shifted spearman correlation on the individual data shifted
to the interval [0, 1].
}

