% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ys1_yr1.R
\name{yr1.df}
\alias{yr1.df}
\title{yr1 correlation matrix for given data frame.}
\usage{
yr1.df(data, time_pts, w1 = 0.5, w2 = 0.25, w3 = 0.25,
  use = "pairwise.complete.obs")
}
\description{
This calculates the yr1 correlation matrix for the given set of factors.
}

