% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ys1_yr1.R
\name{ys1.df}
\alias{ys1.df}
\title{Calculate ys1 matrix for a given data frame.}
\usage{
ys1.df(data, time_pts, w1 = 0.5, w2 = 0.25, w3 = 0.25,
  use = "pairwise.complete.obs")
}
\description{
Every column of the data.frame is a single measurement (or mean) of a single factor for
the provided time points.
Use this function to calculate the correlation matrix on the mean data.
TODO: refactor so that only one function for the calculation of correlation matrix exists.
}

