% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ys1_yr1.R
\name{ys1}
\alias{ys1}
\title{ys1 correlation measure.}
\usage{
ys1(a, b, time_pts, w1 = 0.5, w2 = 0.25, w3 = 0.25, use = "all.obs")
}
\description{
the time course vectors for factor a and b and the time_pts are required,
with all 3 vectors having the same length.
}

