% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ys1_yr1.R
\name{ys1.timepoints.df}
\alias{ys1.timepoints.df}
\title{Calculate ys1 for factors provided as list of matrices with size [Nt, Nr].}
\usage{
ys1.timepoints.df(data_list, time_pts, w1 = 0.5, w2 = 0.25, w3 = 0.25,
  use = "pairwise.complete.obs")
}
\description{
Calculate ys1 for factors provided as list of matrices with size [Nt, Nr].
}

