% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ys1_yr1.R
\name{ys1.timepoints}
\alias{ys1.timepoints}
\title{ys1 calculation for timepoint data.}
\usage{
ys1.timepoints(a, b, time_pts, w1 = 0.5, w2 = 0.5, use = "all.obs")
}
\description{
Every measurement belongs to one sample. The columns are the number of repeats.
}

