% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ys1_yr1.R
\name{ys2.df}
\alias{ys2.df}
\title{ys2 correlation matrix.}
\usage{
ys2.df(data, time_pts, w1 = 0.5, w2 = 0.25, w3 = 0.25,
  use = "pairwise.complete.obs")
}
\description{
TODO: refactor in one function
}

