% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ys1_yr1.R
\name{ysr}
\alias{ysr}
\title{YS and YR correlation measure.}
\usage{
ysr(a, b, time_pts, use = "all.obs")
}
\description{
The time course vectors for factor a and b and the time_pts are required,
with all 3 vectors having the same length.
This function only calculates the components of the various correlation
scores which still have to be combined via respective weights for the
components, i.e. for instance
w1=0.50, w2=0.25, w3=0.25
}

