% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_reshape.R
\name{bdl_mean_data}
\alias{bdl_mean_data}
\title{Calculation of mean data based on provided time factor.}
\usage{
bdl_mean_data(data)
}
\description{
Calculation of mean data based on provided time factor.
}

