% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BDLanalysis.R
\docType{data}
\name{BDLdata}
\alias{BDLdata}
\title{Time course data from BDL mice.}
\format{A data frame with with 40 rows and 154 variables: \code{Ppara}, \code{Cyp3a11},
  \code{Cyp24a1}, ...).}
\description{
Complete data set of all time course data for all factors. The sample
description is available in BDLsamples.
}

