% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BDLdata}
\alias{BDLdata}
\title{Time course factor data for BDL mice.}
\format{A data frame with factor variables: \code{Ppara}, \code{Cyp3a11},
  \code{Cyp24a1}, ...).}
\usage{
BDLdata
}
\description{
Complete data set of all time course data for all factors. The sample
description is available in BDLsamples.
}
\keyword{datasets}

