% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BDLanalysis.R
\docType{data}
\name{BDLsamples}
\alias{BDLsamples}
\title{Sample description for BDLdata.}
\format{A data frame with 40 rows and 9 variable providing additional information for the samples: \code{time}, \code{time_fac}, \code{time_point}, \code{repeats} ...).}
\description{
Sample description for the BDLdata factor data set. Here the time point and
repeat information is defined.
}

