% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BDLsamples}
\alias{BDLsamples}
\title{Sample description corresponding to BDLdata.}
\format{A data frame with additional information for the samples: \code{time}, \code{time_fac}, \code{time_point}, \code{repeats} ...).}
\usage{
BDLsamples
}
\description{
Sample description for the BDLdata factor data set. Here the time point and
repeat information is defined.
}
\keyword{datasets}

