% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_single_factor}
\alias{plot_single_factor}
\title{Create a single factor plot.}
\usage{
plot_single_factor(name)
}
\description{
Plot the time course of a single factor, i.e. all individual data points.
}

